+++ vector/src/main/java/im/vector/riotx/features/login/LoginWebFragment.kt
@@ -21,6 +21,7 @@ package im.vector.riotx.features.login
 import android.annotation.SuppressLint
 import android.content.DialogInterface
 import android.graphics.Bitmap
+import android.net.Uri
 import android.net.http.SslError
 import android.os.Build
 import android.os.Bundle
@@ -36,6 +37,7 @@ import im.vector.matrix.android.api.auth.REGISTER_FALLBACK_PATH
 import im.vector.matrix.android.api.auth.SSO_FALLBACK_PATH
 import im.vector.matrix.android.api.auth.SSO_REDIRECT_URL_PARAM
 import im.vector.matrix.android.api.auth.data.Credentials
+import im.vector.matrix.android.api.extensions.tryThis
 import im.vector.matrix.android.internal.di.MoshiProvider
 import im.vector.riotx.R
 import im.vector.riotx.core.extensions.appendParamToUrl
@@ -54,6 +56,11 @@ import javax.inject.Inject
 class LoginWebFragment @Inject constructor(
         private val assetReader: AssetReader
 ) : AbstractLoginFragment() {
+
+    companion object {
+        // Note that the domain can be displayed to the user for confirmation that he trusts it. So use a human readable string
+        private const val REDIRECT_URL = "riotx://riotx"
+    }
 
     override fun getLayoutResId() = R.layout.fragment_login_web
 
@@ -130,12 +137,8 @@ class LoginWebFragment @Inject constructor(
             if (state.signMode == SignMode.SignIn) {
                 if (state.loginMode == LoginMode.Sso) {
                     append(SSO_FALLBACK_PATH)
-                    // We do not want to deal with the result, so let the fallback login page to handle it for us
-                    appendParamToUrl(SSO_REDIRECT_URL_PARAM,
-                            buildString {
-                                append(state.homeServerUrl?.trim { it == '/' })
-                                append(LOGIN_FALLBACK_PATH)
-                            })
+                    // Set a redirect url we will intercept later
+                    appendParamToUrl(SSO_REDIRECT_URL_PARAM, REDIRECT_URL)
                 } else {
                     append(LOGIN_FALLBACK_PATH)
                 }
@@ -226,7 +229,9 @@ class LoginWebFragment @Inject constructor(
              * @return
              */
             override fun shouldOverrideUrlLoading(view: WebView, url: String?): Boolean {
-                if (null != url && url.startsWith("js:")) {
+                if (url == null) return super.shouldOverrideUrlLoading(view, url as String?)
+
+                if (url.startsWith("js:")) {
                     var json = url.substring(3)
                     var javascriptResponse: JavascriptResponse? = null
 
@@ -256,11 +261,21 @@ class LoginWebFragment @Inject constructor(
                         }
                     }
                     return true
+                } else if (url.startsWith(REDIRECT_URL)) {
+                    return handleSsoLoginSuccess(url)
                 }
 
                 return super.shouldOverrideUrlLoading(view, url)
             }
         }
+    }
+
+    private fun handleSsoLoginSuccess(url: String): Boolean {
+        val uri = Uri.parse(url)
+        val loginToken = tryThis { uri.getQueryParameter("loginToken") } ?: return false
+
+        loginViewModel.handle(LoginAction.LoginWithToken(loginToken))
+        return true
     }
 
     private fun notifyViewModel(credentials: Credentials) {